+++ mopidy/http/actor.py
@@ -19,7 +19,7 @@ from mopidy.http import Extension, handlers
 from mopidy.internal import formatting, network
 
 if TYPE_CHECKING:
-    from typing import Any, ClassVar, List, Type
+    from typing import Any, ClassVar
 
 try:
     import asyncio
@@ -31,8 +31,8 @@ logger = logging.getLogger(__name__)
 
 
 class HttpFrontend(pykka.ThreadingActor, CoreListener):
-    apps: ClassVar[List[Type[Any]]] = []
-    statics: ClassVar[List[Type[Any]]] = []
+    apps: ClassVar[list[type[Any]]] = []
+    statics: ClassVar[list[type[Any]]] = []
 
     def __init__(self, config, core):
         super().__init__()